debxo: drop gnome ubifs images
Gnome barely fits in the 1GB XO-1 nand; when running dist-upgrade
inside of gnome, gconftool-2 easily eats up all 256MB of ram and
causes an OOM crash.  The gnome utilities quickly exhaust the XO-1s
ram, as well.  As such, gnome will no longer be supported for the
XO-1 nand.  Drop the ubifs config files accordingly.

It will remain supported for the XO-1.5, as well as larger SD/USB
devices on XO-1.  Don't expect me to care about memory usage,
though! :)

+++ release.sh
@@ -8,12 +8,12 @@ for f in configs/debxo-*; do
 deb http://localhost:9999/security squeeze/updates main contrib non-free
 deb http://localhost:9999/backports squeeze-backports main contrib non-free" \
 			debxo-${desktop}
-	./mkubifs.sh --config-type debxo-${desktop} debxo-${desktop} ${desktop}.img
-	./mkext3.sh --config-type debxo-${desktop} debxo-${desktop} \
+	[ -f ./configs/debxo-${desktop}/fstab-ubifs ] && ./mkubifs.sh --config-type debxo-${desktop} debxo-${desktop} ${desktop}.img
+	[ -f ./configs/debxo-${desktop}/fstab-ext3 ] && ./mkext3.sh --config-type debxo-${desktop} debxo-${desktop} \
 			debxo-${desktop}.ext3.img
-	gzip debxo-${desktop}.ext3.img
+	[ -f debxo-${desktop}.ext3.img ] && gzip debxo-${desktop}.ext3.img
 
 	mkdir -p ext3 nand
-	mv debxo-${desktop}.ext3.img.gz ext3
-	mv ${desktop}.img ${desktop}.dat nand
+	[ -f debxo-${desktop}.ext3.img.gz ] && mv debxo-${desktop}.ext3.img.gz ext3
+	[ -f ${desktop}.img ] && mv ${desktop}.img ${desktop}.dat nand
 done
